Move ProminentWord value to a new folder inside languageProcessing
