+++ packages/yoastseo/spec/languageProcessing/helpers/prominentWords/determineProminentWordsSpec.js
@@ -1,4 +1,4 @@
-import ProminentWord from "../../../../src/values/ProminentWord";
+import ProminentWord from "../../../../src/languageProcessing/values/ProminentWord";
 import {
 	getProminentWords,
 	getProminentWordsFromPaperAttributes,
