+++ packages/yoastseo/src/values/index.js
@@ -2,7 +2,7 @@ import AssessmentResult from "./AssessmentResult";
 import Mark from "./Mark";
 import Paper from "./Paper";
 import Participle from "./Participle";
-import ProminentWord from "./ProminentWord";
+import ProminentWord from "../languageProcessing/values/ProminentWord";
 import Sentence from "./Sentence";
 import SentencePart from "./SentencePart";
